+++ src/config/mod.rs
@@ -512,6 +512,9 @@ pub struct Config {
     pub launch_menu: Vec<SpawnCommand>,
 
     #[serde(default = "default_true")]
+    pub add_wsl_distributions_to_launch_menu: bool,
+
+    #[serde(default = "default_true")]
     pub check_for_updates: bool,
 
     #[serde(default = "default_update_interval")]